+++ src/NetAppFiles/NetAppFiles/Account/NewNetAppFilesAccount.cs
@@ -84,7 +84,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Account
             var netAppAccountBody = new NetAppAccount()
             {
                 Location = Location,
-                ActiveDirectories = (ActiveDirectory != null) ? ModelExtensions.ConvertActiveDirectoriesFromPs(ActiveDirectory) : null,
+                ActiveDirectories = (ActiveDirectory != null) ? ActiveDirectory.ConvertFromPs() : null,
                 Tags = tagPairs
             };