+++ src/EventHub/EventHub/ChangeLog.md
@@ -18,8 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 1.7.1
 * Fixed Cluster commands for EventHub cluster without tags
-* updated help text for PartnerNamespace of AzEventHubGeoDRConfiguration commands 
+* Updated help text for PartnerNamespace of AzEventHubGeoDRConfiguration commands 
 
 ## Version 1.7.0
 * Added optional switch parameter `TrustedServiceAccessEnabled` to `Set-AzEventHubNetworkRuleSet` cmdlet